+++ src/features/schedule/pages/Calendars.jsx
@@ -1,248 +1,41 @@
-import { useEffect, useMemo, useRef, useState } from "react";
-import FullCalendar from "@fullcalendar/react";
-import dayGridPlugin from "@fullcalendar/daygrid";
-import interactionPlugin from "@fullcalendar/interaction";
-import listPlugin from "@fullcalendar/list";
-import koLocale from "@fullcalendar/core/locales/ko";
-import styles from "../styles/Calendars.module.css";
+import { useMemo } from "react";
+import styles from "../styles/CalendarShell.module.css";
+
+import CalendarShell from "../components/calendar/CalendarShell";
+import { normalizeEvents } from "../../../utill/calendar/normalizeEvents";
+
 import data from "../../../utill/date.js";
 import albums from "../../../utill/albumdata.js";
 import concerts from "../../../utill/concertdata.js";
-import WeekPager from "../components/WeekPager.jsx";
-
-function useIsMobile() {
-  const [isMobile, setIsMobile] = useState(
-    typeof window !== "undefined" &&
-      window.matchMedia("(max-width: 1024px)").matches
-  );
-  useEffect(() => {
-    const mq = window.matchMedia("(max-width: 1024px)");
-    const onChange = (e) => setIsMobile(e.matches);
-    mq.addEventListener("change", onChange);
-    return () => mq.removeEventListener("change", onChange);
-  }, []);
-  return isMobile;
-}
 
 export default function Calendars({ artist, album, concert }) {
-  const calendarRef = useRef(null);
-  const monthRef = useRef(null);
-  const listRef = useRef(null);
-  const isMobile = useIsMobile();
-  const [listAnchor, setListAnchor] = useState(null);
-  const [listBase, setListBase] = useState(null);
-  const monthWrapRef = useRef(null);
-  const listWrapRef  = useRef(null);
-  const koNoAllDay = useMemo(() => ({ ...koLocale, allDayText: "" }), []);
-  const [monthRange, setMonthRange] = useState({ start: null, end: null });
-  
-  // 한 곳에서 이벤트 합치기
-  const allEvents = useMemo(() => {
-    const ymdLocal = (d) => {
-      const dt = d instanceof Date ? d : new Date(d);
-      if (Number.isNaN(dt.getTime())) return null;
-      const y = dt.getFullYear();
-      const m = String(dt.getMonth() + 1).padStart(2, "0");
-      const da = String(dt.getDate()).padStart(2, "0");
-      return `${y}-${m}-${da}`;          // ← 로컬 기준
-    };
-    const normalize = (val) => {
-      if (!val) return null;
-      if (typeof val === "string") {
-        if (/^\d{4}-\d{2}-\d{2}$/.test(val)) return val; // 이미 YYYY-MM-DD
-        return val.split("T")[0];                        // '2025-10-11T..' → '2025-10-11'
-      }
-      return ymdLocal(val);                              // Date 객체
-    };
-
-    const merged = [...data(artist), ...albums(album), ...concerts(concert)]
-      .map((ev) => {
-        const start = normalize(ev.start ?? ev.date ?? ev.startDate ?? ev.when);
-        const end   = normalize(ev.end ?? ev.endDate ?? null);
-        return {
-          ...ev,
-          title: (ev.title || "").replace(/\s+/g, " ").trim(),
-          start,
-          ...(end ? { end } : {}),
-          allDay: ev.allDay ?? true,
-          extendedProps: { ...(ev.extendedProps || {}), type: ev.type },
-        };
-      })
-      .filter((ev) => !!ev.start);
-
-    // 중복 제거(선택)
-    const seen = new Set();
-    const uniq = [];
-    for (const ev of merged) {
-      const key = `${ev.title}|${ev.extendedProps?.type ?? ""}|${ev.start}|${ev.end ?? ""}`;
-      if (!seen.has(key)) { seen.add(key); uniq.push(ev); }
-    }
-    return uniq;
+  // 메인 페이지용 이벤트 구성
+  const events = useMemo(() => {
+    const raw = [
+      ...(data(artist) || []),
+      ...(albums(album) || []),
+      ...(concerts(concert) || []),
+    ];
+    return normalizeEvents(raw);
   }, [artist, album, concert]);
 
-  const monthFilteredEvents = useMemo(() => {
-    if (!monthRange.start || !monthRange.end) return [];
-    return allEvents.filter((ev) => {
-      const s = new Date(ev.start);
-      return s >= monthRange.start && s < monthRange.end;
-    });
-    // 달과 '겹치는' 모든 이벤트를 넣고 싶으면 아래로 교체
-    // return allEvents.filter((ev) => {
-    //   const s = new Date(ev.start);
-    //   const e = new Date(ev.end || ev.start);
-    //   return s < monthRange.end && e >= monthRange.start;
-    // });
-  }, [allEvents, monthRange]);
-
-  // 월간 뷰의 anchor(해당 달 시작일) 보관 → WeekPager 계산에 사용
-  const [monthAnchor, setMonthAnchor] = useState(new Date());
-
-  useEffect(() => {
-    // 헤더/버튼 톤(필요 없으면 제거 가능)
-    const headerEl =
-      calendarRef.current?.querySelector(".fc-header-toolbar");
-    if (headerEl) {
-      headerEl.style.backgroundColor = "#00B6F0";
-      headerEl.style.color = "#1A1A1A";
-    }
-    const buttons = calendarRef.current?.querySelectorAll(".fc-button");
-    buttons?.forEach((btn) => {
-      btn.style.backgroundColor = "#00B6F0";
-      btn.style.borderColor = "#00B6F0";
-      btn.style.color = "#1A1A1A";
-    });
-  }, []);
-
-  /** 리스트를 특정 날짜가 포함된 주로 이동 */
-  const goListTo = (date) => {
-    const api = listRef.current?.getApi();
-    if (!api) return;
-    api.gotoDate(date);
-    setListAnchor(new Date(date));
-  };
-
-  const skipOnceRef = useRef(false);
-
-  // ✅ 월간(prev/next/today 등) 이동 시 호출
-  const onMonthDatesSet = () => {
-    // 제목에 표시된 '그 달'을 대표하는 날짜
-    const anchorDate = monthRef.current?.getApi()?.getDate() ?? new Date();
-
-    // WeekPager 계산용 anchor도 이걸 쓰기
-    setMonthAnchor(anchorDate);
-
-    // 현재 달 범위
-    const y = anchorDate.getFullYear();
-    const m = anchorDate.getMonth();
-    const monthStart = new Date(y, m, 1);       // inclusive
-    const monthEnd   = new Date(y, m + 1, 1);   // exclusive
-    setMonthRange({ start: monthStart, end: monthEnd });
-
-    // 리스트 시작은 오늘이 그 달이면 오늘, 아니면 그 달 1일
-    const today = new Date();
-
-    // allEvents에서 해당 달의 가장 이른 일정일
-    const firstEventDate =
-      allEvents
-        .map(ev => new Date(ev.start))
-        .filter(d => d >= monthStart && d < monthEnd)
-        .sort((a, b) => a - b)[0] || null;
-
-
-    // 오늘이 그 달이면 오늘, 아니면 첫 일정일(없으면 1일)
-    const base = (today >= monthStart && today < monthEnd)
-      ? today
-      : (firstEventDate || monthStart);
-    setListBase(base);
-    const api = listRef.current?.getApi();
-    api?.gotoDate(base);
-    setListAnchor(new Date(base));
-    skipOnceRef.current = true;
-  };
-
-  useEffect(() => {
-    const m = monthWrapRef.current;
-    const l = listWrapRef.current;
-    if (!m || !l) return;
-
-    const sync = () => {
-      const h = m.clientHeight || 0;
-      l.style.height = `${h}px`;
-    };
-
-    // ResizeObserver가 있으면 사용, 없으면 리사이즈 fallback
-    const RO = window.ResizeObserver;
-    let ro;
-    if (RO) {
-      ro = new RO(sync);
-      ro.observe(m);
-    } else {
-      window.addEventListener("resize", sync);
-    }
-    sync();
-
-    return () => {
-      if (ro) ro.disconnect();
-      else window.removeEventListener("resize", sync);
-    };
-  }, [isMobile]);
-
-  useEffect(() => {
-    if (!monthRange.start || !monthRange.end || !listRef.current) return;
-    
-    if (skipOnceRef.current) {    // 첫 실행은 건너뛰기
-      skipOnceRef.current = false;
-      return;
-    }
-
-    const api = listRef.current.getApi();
-    const viewStart = api.view.currentStart; // 현재 리스트 주 시작
-    const viewEnd   = api.view.currentEnd;   // 현재 리스트 주 끝(다음 주 시작)
-
-    const hasThisWeek = monthFilteredEvents.some(ev => {
-      const s = new Date(ev.start);
-      const e = new Date(ev.end || ev.start);
-      return s < viewEnd && e >= viewStart; // 주와 겹치면 일정 있음
-    });
-
-    if (!hasThisWeek) {
-      const monthEvs = monthFilteredEvents
-        .map(ev => ({ d: new Date(ev.start), ev }))
-        .sort((a, b) => a.d - b.d);
-
-      // 현재 주 이후 가장 가까운 일정(없으면 이전으로)
-      const next =
-        monthEvs.find(x => x.d >= viewEnd)?.d ||
-        [...monthEvs].reverse().find(x => x.d < viewStart)?.d;
-
-      if (next) {
-        api.gotoDate(next);
-        setListAnchor(new Date(next));
-      }
-    }
-  }, [monthRange, monthFilteredEvents]);
-
+  // 색상 범례(지금 너가 쓰던 로직 유지)
   const colorLegend = useMemo(() => {
     const map = new Map();
 
-    // artist: 그룹 색상
     (artist || []).forEach((a) => {
       if (a?.group && a?.color) map.set(a.group, a.color);
     });
 
-    // concert: 그룹 색상
     (concert || []).forEach((c) => {
       if (Array.isArray(c.concertdate)) {
         c.concertdate.forEach((cd) => {
-          const key = cd.artistname || c.group;  // 세부 유닛 우선, 없으면 상위 그룹
-          if (key && cd.color) {
-            map.set(key, cd.color);
-          }
+          const key = cd.artistname || c.group;
+          if (key && cd.color) map.set(key, cd.color);
         });
       }
     });
 
-    // album: 곡 주체(솔로/그룹) 색상
     (album || []).forEach((al) => {
       (al.music || []).forEach((m) => {
         const key = m.groupsolo || al.group || m.albumname;
@@ -250,7 +43,6 @@ export default function Calendars({ artist, album, concert }) {
       });
     });
 
-    // [{label, color}] 배열로 변환
     return Array.from(map, ([label, color]) => ({ label, color }));
   }, [artist, album, concert]);
 
@@ -262,188 +54,22 @@ export default function Calendars({ artist, album, concert }) {
     { code: "T", label: "Etc / 기타" },
   ];
 
+  // 메인 페이지 테마(기존 #00B6F0 유지)
+  const theme = {
+    headerBg: "#00B6F0",
+    headerText: "#1A1A1A",
+    buttonBg: "#00B6F0",
+    buttonText: "#1A1A1A",
+  };
 
   return (
-    <section className={styles.CalendarSection}>
-
-      <div ref={calendarRef} className={styles.Calendars}>
-        {/* 월간 */}
-        <div ref={monthWrapRef} className={styles.month}>
-          <FullCalendar
-            plugins={[dayGridPlugin, interactionPlugin]}
-            ref={monthRef}
-            locale={koLocale}
-            initialView="dayGridMonth"
-            headerToolbar={{ left: "prev today", center: "title", right: "next" }}
-            editable
-            selectable
-            selectMirror
-            timeZone="local"
-            dayMaxEvents={false}
-            fixedWeekCount={false}        // 6주 고정 해제 → 4~6주 가변
-            showNonCurrentDates={false}   // 이번 달 아닌 날짜는 셀 자체를 숨김
-            height={isMobile ? "auto" : "auto"}
-            contentHeight="auto"
-            expandRows={true}
-            firstDay={0} 
-            datesSet={onMonthDatesSet}
-            dateClick={(info) => goListTo(info.date)} // 날짜 탭 → 해당 주로 리스트 이동
-            events={[]}
-            views={{
-              dayGridMonth: { eventDisplay: 'none', displayEventTime: false }
-            }}
-            dayCellDidMount={(info) => {
-              // 비월 셀은 FullCalendar가 클래스로 구분 → 이것만 제외
-              if (info.el.classList.contains("fc-day-other")) return;
-
-              // 중복 wrap 제거
-              const frame = info.el.querySelector(".fc-daygrid-day-frame");
-              if (!frame) return;
-              frame.style.position = "relative";
-              frame.querySelector(".fc-dots-wrap")?.remove();
-
-              // ✅ 수정: ISO 날짜 문자열로만 정확히 비교 (타임존 영향 배제)
-              const ymdLocal = (d) => {
-                const y = d.getFullYear();
-                const m = String(d.getMonth() + 1).padStart(2, "0");
-                const da = String(d.getDate()).padStart(2, "0");
-                return `${y}-${m}-${da}`;
-              };
-              const cellISO = ymdLocal(info.date);
-              const dayEvents = allEvents.filter((ev) => {
-                const s = ev.start;        // 이미 YYYY-MM-DD로 정규화되어 있음
-                const e = ev.end || s;     // 종일 단일 이벤트
-                // s <= cellISO <= e  (exclusive-end가 아니라 inclusive로 처리)
-                return (s <= cellISO) && (cellISO <= e);
-              });
-
-              // 점 컨테이너 (날짜 숫자 아래에 세로로 쌓이게)
-              const wrap = document.createElement("div");
-              wrap.className = "fc-dots-wrap";
-              Object.assign(wrap.style, {
-                position: "absolute",
-                top: "50%",
-                left: "10px",
-                transform: "translateY(-50%)",
-                display: "flex",
-                flexDirection: "column",
-                gap: "6px",
-                pointerEvents: "none",
-                zIndex: 2
-              });
-
-              // 점 크기/개수 여기서 조절
-              const DOT_SIZE = 12;   // ← 점 크기
-              const MAX_DOTS = 8;    // ← 최대 개수
-
-              dayEvents.slice(0, MAX_DOTS).forEach((ev) => {
-                const dot = document.createElement("span");
-                dot.className = "fc-dot";
-                const color = ev.color || ev.backgroundColor || "#00B6F0";
-                const type  = ev.extendedProps?.type || ev.type;
-
-                dot.style.cssText = `
-                  width:${DOT_SIZE}px;height:${DOT_SIZE}px;border-radius:50%;
-                  display:inline-block;flex:0 0 ${DOT_SIZE}px;background:${color};
-                  box-sizing:border-box;
-                `;
-
-                wrap.appendChild(dot);
-              });
-
-              frame.appendChild(wrap);
-            }}
-          />
-        </div>
-
-        {/* 리스트(이번 주만) + 주별 페이지네이션 */}
-        <div ref={listWrapRef} className={styles.list}>
-          <FullCalendar
-            plugins={[listPlugin]}
-            key={listBase?.toISOString() || monthRange.start?.toISOString() || "list-init"}
-            ref={listRef}
-            locale={koNoAllDay}
-            initialView="listWeek"
-            initialDate={listBase || monthRange.start || new Date()}
-            headerToolbar={false}
-            height="100%"
-            firstDay={0} 
-            noEventsContent={<div className={styles.noEventsBox}>일정이 없습니다</div>}
-            timeZone="local"
-            listDayFormat={{ weekday: "short" }} // WED
-            listDaySideFormat={{ day: "2-digit" }} // 23
-            events={monthFilteredEvents}
-            eventContent={({ event }) => {
-              const type = event.extendedProps?.type; // ETC 기본값 제거
-              const wrap = document.createElement("div");
-              wrap.className = styles.evRow;
-
-              if (type) {
-                const badge = document.createElement("span");
-                badge.className = `${styles.badge} ${styles.badgeLg} ${styles["type"+type]}`;
-                badge.textContent = type;
-                wrap.appendChild(badge);
-              }
-
-              const title = document.createElement("span");
-              title.className = styles.evTitle;
-              title.textContent = event.title;
-              wrap.appendChild(title);
-
-              return { domNodes: [wrap] };
-            }}
-          />
-
-          {/* 주별 페이지네이션 (해당 달 한정) */}
-          {listAnchor && (
-              <WeekPager
-                key={monthAnchor?.toISOString() + '|' + listAnchor?.toISOString()} // 선택: 초기 깜빡임 방지
-                monthAnchor={monthAnchor}
-                onGotoWeek={(date) => {
-                  const api = listRef.current?.getApi();
-                  api?.gotoDate(date);
-                  setListAnchor(new Date(date));
-                }}
-                currentDate={listAnchor}
-              />
-            )}
-        </div>
-        {/* === 범례 === */}
-        <div className={styles.legendWrap}>
-          <div className={styles.legendBlock}>
-            <div className={styles.legendTitle}>유형</div>
-            <div className={styles.typeLegendRow}>
-              {typeLegend.map((t) => (
-                <div key={t.code} className={styles.typeLegendItem}>
-                  <span className={`${styles.badge} ${styles["type" + t.code]}`}>
-                    {t.code}
-                  </span>
-                  <span className={styles.legendText}>{t.label}</span>
-                </div>
-              ))}
-            </div>
-          </div>
-
-          <div className={styles.legendBlock}>
-            <div className={styles.legendTitle}>색상</div>
-            <div className={styles.colorLegendGrid}>
-              {colorLegend.map((c) => (
-                <div key={c.label} className={styles.colorLegendItem}>
-                  <span
-                    className={styles.colorDot}
-                    style={{ background: c.color }}
-                    aria-label={c.label}
-                  />
-                  <span className={styles.legendText}>{c.label}</span>
-                </div>
-              ))}
-              {colorLegend.length === 0 && (
-                <div className={styles.legendEmpty}>표시할 항목이 없어요</div>
-              )}
-            </div>
-          </div>
-        </div>
-      </div>
-    </section>
+    <CalendarShell
+      styles={styles}
+      events={events}
+      colorLegend={colorLegend}
+      typeLegend={typeLegend}
+      theme={theme}
+      monthOnly={false}
+    />
   );
 }
